backend/requirements.txt
browser-use>=0.12.6
google-genai
fastapi
uvicorn[standard]
python-dotenv
